gestion des temps

+++ app.py
@@ -8,11 +8,13 @@ from apscheduler.schedulers.background import BackgroundScheduler
 from apscheduler.triggers.interval import IntervalTrigger
 from slackclient import SlackClient
 
-
 app = Flask(__name__)
 
-
 app.config.from_object('secret_config')
+
+# --------------------------------------------------
+# ------------------- Functions --------------------
+# --------------------------------------------------
 
 def connect_db () :
     g.mysql_connection = mysql.connector.connect(
@@ -47,13 +49,6 @@ def getHttpCode (url) :
     except urllib.error.URLError as error :
         return error.reason
 
-def isUrlCorrect (url) :
-    try :
-        request = urllib.request.urlopen(url)
-        return True
-    except urllib.error.URLError as error :
-        return error.reason
-
 def getStatus(sites) :
     status = []
     for site in sites :
@@ -61,41 +56,44 @@ def getStatus(sites) :
     return status
 
 def insertStatusCode () :
-    db = get_db()
-    db.execute("SELECT id, url FROM sites")
-    sites = db.fetchall()
-    date = datetime.datetime.now()
-    datas = []
-
-    for site in sites:
-        code = getHttpCode(site[1])
-        datas.append((site[0], code, date))
-
-    sql = "INSERT INTO status VALUES(DEFAULT, %s, %s, %s)"
-    db.executemany(sql, datas)
-    g.mysql_connection.commit()
-
-def autoInsertStatusCode () :
     db = newConnectDb()
     cursor = db.cursor()
     cursor.execute("SELECT id, url FROM sites")
     sites = cursor.fetchall()
     date = datetime.datetime.now()
-    datas = []
 
     for site in sites:
         code = getHttpCode(site[1])
-        if code != 200 :
-            if isDown(site[0]) :
-                sendSlackMessage(site[1], code)
-                sendTelegramMessage(site[1], code)
-        datas.append((site[0], code, date))
-
-    sql = "INSERT INTO status VALUES(DEFAULT, %s, %s, %s)"
-    cursor.executemany(sql, datas)
-    db.commit()
+        sql = "INSERT INTO status VALUES(DEFAULT, %s, %s, %s)"
+        cursor.execute(sql, (site[0], code, date))
+        db.commit()
+        if code != 200 and isDown(site[0], 3) :
+            sendMessages(site[0], code)
     cursor.close()
     db.close()
+
+def sendMessages(siteId, code, antiSpam = True) :
+    if antiSpam :
+        db = newConnectDb()
+        cursor = db.cursor()
+        cursor.execute("SELECT * FROM sites WHERE id = %(siteId)s", ({'siteId': siteId}))
+        data = cursor.fetchone()
+        now = datetime.datetime.now()
+        if data[2] != None :
+            diff = now - data[2]
+            days, seconds = diff.days, diff.seconds
+            hours = days * 24 + seconds // 3600
+            minutes = (seconds % 3600) // 60
+        if data[2] == None or hours >= 2 :
+            sendSlackMessage(data[1], code)
+            sendTelegramMessage(data[1], code)
+            cursor.execute("UPDATE sites SET last_message = %(date)s WHERE id = %(id)s", ({'date': now, 'id': siteId}))
+            db.commit()
+        cursor.close()
+        db.close()
+    else :
+        sendSlackMessage(data[1], code)
+        sendTelegramMessage(data[1], code)
 
 def sendSlackMessage (url, code) :
     slack_token = app.config['SLACK_TOKEN']
@@ -111,16 +109,17 @@ def sendTelegramMessage (url, code) :
     bot = telegram.Bot(token=telegram_token)
     bot.send_message(chat_id=app.config['TELEGRAM_CHAT_ID'], text="Bonjour, il semblerait qu'il y ai un problème avec votre site "+ url +". Statut : "+ str(code))
 
-def isDown (id) :
+def isDown (siteId, limit) :
     db = newConnectDb()
     cursor = db.cursor()
-    cursor.execute("SELECT * FROM status WHERE site_id = %(id)s ORDER BY date DESC LIMIT 3", {'id' : id})
+    cursor.execute("SELECT * FROM status WHERE site_id = %(id)s ORDER BY date DESC LIMIT %(limit)s", {'id': siteId, 'limit': limit})
     lastStatus = cursor.fetchall()
+    cursor.close()
+    db.close()
     for status in lastStatus :
-        if status[2] != 200 :
+        if status[2] == 200 :
             return False
     return True
-
 
 # Ferme les BDD à la fin de l'execution
 @app.teardown_appcontext
@@ -128,8 +127,9 @@ def close_db (error) :
     if hasattr(g, 'db') :
         g.db.close()
 
-
-# Routes
+# --------------------------------------------------
+# ---------------------- Routes --------------------
+# --------------------------------------------------
 
 @app.route('/')
 def index () :
@@ -179,9 +179,6 @@ def admin () :
     db.execute("SELECT id, url FROM sites")
     sites = db.fetchall()
     status = getStatus(sites)
-    # insertStatusCode()
-    # sendSlackMessage()
-    # sendTelegramMessage()
 
     return render_template('admin.html.j2', user = session['user'], sites = sites, status = status)
 
@@ -195,7 +192,7 @@ def add () :
     if url is not None:
         db = get_db()
         if type(getHttpCode(url)) is int :
-            db.execute("INSERT INTO sites(id, url) VALUES(DEFAULT, '%s')"%(str(url)))
+            db.execute("INSERT INTO sites(id, url, last_message) VALUES(DEFAULT, '%s', NULL)"%(str(url)))
             g.mysql_connection.commit()
             flash("Le site " + url + " a bien été ajouté à la liste")
             return redirect(url_for('admin'))
@@ -221,7 +218,7 @@ def edit (id) :
             flash("Ce site ne semble pas exister")
             return redirect(url_for('admin'))
     else:
-        if (isUrlCorrect(url) == True) :
+        if type(getHttpCode(url)) is int :
             db.execute("UPDATE sites SET url = '%s' WHERE id = %s"%(str(url), id))
             g.mysql_connection.commit()
             flash("Le site ayant l'id " + str(id) + " a bien été modifié avec la valeur " + url)
@@ -230,21 +227,30 @@ def edit (id) :
             flash("Le site " + url + " ne semble pas exister ou n'autorise pas que vous l'analysiez")
             return render_template('edit.html.j2', site = site[0], user = session['user'])
 
-@app.route('/admin/delete/<int:id>')
+@app.route('/admin/delete/<int:id>', methods = ['GET', 'POST'])
 def delete (id) :
     if not session.get('user') or not session.get('user')[1] :
         return redirect(url_for('login'))
 
     db = get_db()
-    db.execute("DELETE FROM sites WHERE id = %s"%(id))
-    g.mysql_connection.commit()
+    db.execute("SELECT * FROM sites WHERE id = %(id)s", ({'id': id}))
+    db.fetchone()
 
     if db.rowcount > 0 :
-        flash("Le site ayant l'id " + str(id) + " a bien été supprimé")
-    else :
-        flash("Le site ayant l'id " + str(id) + " n'a pas pu être supprimé")
+        if request.method == 'POST' :
+            db.execute("DELETE FROM sites WHERE id = %s"%(id))
+            g.mysql_connection.commit()
 
-    return redirect(url_for('admin'))
+            if db.rowcount > 0 :
+                flash("Le site ayant l'id " + str(id) + " a bien été supprimé")
+            else :
+                flash("Le site ayant l'id " + str(id) + " n'a pas pu être supprimé")
+            return redirect(url_for('admin'))
+        else :
+            return render_template('delete.html.j2', user = session['user'])
+    else :
+        flash("Le site ayant l'id " + str(id) + " ne semble pas exister")
+        return redirect(url_for('admin'))
 
 @app.route('/history/<int:id>')
 def history (id) :
@@ -258,15 +264,17 @@ def history (id) :
     status = db.fetchall()
     return render_template('history.html.j2', user = user, status = status)
 
+# Automatise en arrière plan la vérification des statuts et ce qui s'en suit
 with app.app_context():
     if os.environ.get("WERKZEUG_RUN_MAIN") == "true":
         scheduler = BackgroundScheduler()
         scheduler.add_job(
-            func=autoInsertStatusCode,
-            trigger=IntervalTrigger(seconds=10),
+            func=insertStatusCode,
+            trigger=IntervalTrigger(seconds=120),
             replace_existing=True,
         )
         scheduler.start()
+        print(" * Starting scheduler")
 
 if __name__ == '__main__':
     app.run(debug=True, host='0.0.0.0')
